Update tests and OwnerID extraction
Turns out openaai package has an OwnerContext that's packing OwnerID
which is set to the User.ID.

Unfortunately Runners do not have User.ID set, so we are hardcoding the
OwnerID for runners to RunnerID set to `runner` for every runner.
This is already done for `system` apparently, but it's not ideal.

We need to replace the RunnerID with the actual Runner ID in the future
but for the time being this will do.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>