(refactor): Client context plumbing
This commit does helix API context plumbing.
The API client now accepts context which is passed down to it from a
callsite.

In some helix commands we do this, but it others we dont. This is a
terrible inconsistency that might prevent adding some fuctionality in
the future (say, cli canellations, etc.)

makeRequest now accepts context which is then wrapped in the request timeout.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>