feat(runner)!: NATS based runner communication architecture (#812)
This introduces a big re-architecture of the way we communicate with runners. Previously, runners were expected to reach out to the control plane. This meant the scheduler wasn't truly in control over what runners were doing and resulted in a variety of hard to manage race conditions.

Background
Now (with one exception noted below), the runner connects to the runner once to establish a web socket connection, which NATS then uses. All subsequent communication is RPC from the control plane via NATS. The runner is implemented as a simple web server with the usual REST semantics. So from the control plane, it's almost like doing a simple REST request to tell the runner what to do, which makes everything much simpler to manage and will ultimately be far more stable.

Known Problems
The various "upload files" paths to upload results to the control plane are the same.
Image fine tuning wasn't implemented.
(NTH) image generation request doesn't propagate session errors properly, probably need to wire that in like i did text
(NTH) Slots aren't shown on the dash until models are running. But some of them take a while to load, like flux. Would be nice if they were shown immediately with a status.
(NTH) Weird situation when hammering and the hammered nodes have a bit more free GPU left compared to a very stale node. It prefers to try and delete stales on this because there's more room. It's a scheduling strategy improvement.
(NTH) More optimisation could be done on the cold start, since we're limiting to one new slot at a time for safety. Could lock over runner IDs so that it's one start per runner.

BREAKING CHANGE: You must upgrade your runners and control plane to this version at the same time. Neither are backwards compatible.