fix: hide Start Implementation button when spec is approved
When a task's spec is approved (status becomes 'spec_approved'), the
task is already in the implementation phase. The 'Start Implementation'
button should not be visible since implementation has already begun or
is about to begin.

This fix adds the TaskStatusSpecApproved status to the implementationStarted
check in DesignReviewViewer, ensuring the button is hidden when the spec
has been approved.

Fixes #1433

Signed-off-by: majiayu000 <1835304752@qq.com>